requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
beautifulsoup4==4.13.4
bleach==6.2.0
certifi==2025.8.3
cffi==1.17.1
channels==4.3.1
charset-normalizer==3.4.3
constantly==23.10.4
cryptography==45.0.6
daphne==4.2.1
distro==1.9.0
dj-database-url==2.3.0
Django==5.2.5
django-ratelimit==4.1.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jiter==0.10.0
openai==1.99.9
packaging==25.0
pillow==11.3.0
pillow_heif==1.1.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyOpenSSL==25.1.0
python-dotenv==1.1.1
python-magic==0.4.27
requests==2.32.4
service-identity==24.2.0
sniffio==1.3.1
soupsieve==2.7
sqlparse==0.5.3
pytest==8.3.4
pytest-django==4.9.0
tqdm==4.67.1
Twisted==25.5.0
txaio==25.6.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
webencodings==0.5.1
whitenoise==6.9.0
zope.interface==7.2
